Finish replacing Rc<[u8]> with Bytes and BytesMut
This was very hard, because a BytesMut cannot have shared ownership.  Now,
whenever you make an operation using a BytesMut, the Future has exclusive
ownership of the buffer.  It returns the buffer in its Result.

A potential simplification would be for operations on Bytes objects not to
return the buffer, since those buffers support shared ownership.  But for now,
for consistency's sake, we use them the same as BytesMut.

An outstanding bug is that we don't return the correct buffers for writev_at
operations, because the future doesn't know how the various callers' buffers
were combined.