Switch buffer type from bytes to divbuf
The Bytes class seemed like a good idea at the time, but reassembling split
Bytes takes O(n) time.  It's even a little bit worse for RAID operations that
span multiple stripes, because the PRIME-S location function is slightly worse
than O(1).

The DivBuf class can be split just like Bytes, but doing so does not destroy
the reference to the unsplit original.  So "reassembling" is O(1).  In fact,
it probably won't even be necessary to keep a buf element in IoVecResult.