WIP adding a Declust::Locator iterator
An iterator that can iterate through chunk locations would make some of the
VdevRaid programming simpler.  It would also be faster, largely because div
and mod operations could be factored out.

Wrote an implementation of such an iterator for PrimeS.  It runs in 10ns,
compared to 92ns for PrimeS::id2loc.  However, it isn't hooked up to the
Locator Trait yet.