Don't return values we can't compute correctly
The definitions of Vdev::(read|write)_at returned a future containing an
isize; it was supposed to be the amount of data actually written.  This
makes sense for VdevLeaf.  However, it's not always possible to compute it
correctly.  In VdevBlock's case, adjacent op combining makes it impossible
to know which operation was short, if the combined op returns a short value.
We could do it if we kept a lot more information around in the Future, but
the value of that information is dubious anyway.  In VdevRaid's case, it's
impossible to know whether a short write was due to data or parity getting
dropped.  Likewise, the information isn't very useful anyway.

Changing those future's Item types to () entails removing read_at and
write_at from the Vdev trait.  It's ok to do that, since the goal is to have
a fixed vdev hierarchy instead of a flexible one.