Declustered RAID requires an inter-zone gap
Declustered RAID is not monotonic across multiple disks.  That is, while
RAID LBA X may correspond to disk 0 LBA Y, RAID LBA X+1 may correspond to
disk 1 LBA Y-1.  That's bad for zoned devices, if LBA Y begins a zone,
because erasing the zone that begins with Y will also erase part of a stripe
from the previous zone.

To manage this problem, allow for small gaps of unused space between zones.
Basically, any RAID stripe that crosses zones won't be used.  Modify
Vdev::lba2zone to return an Option, and change Vdev::start_of_zone to
Vdev::zone_limits, so the gap will be clear.  Vdev::size's signature will be
unchanged, but the result must not be approximate, because it's impracticle
to calculate the exact amount of space available in all zones whenever that
function is called.

An alternative possibility would've been to restart the declustered layout
at the beginning of each zone, wasting some space at the end of each zone
but not the beginning.  However, that would've required either storing zone
number in the indirect block metadata (yuck), or computing the zone
boundaries on each and every read (yuck again).  This is the least bad
option.