Revamp the I/O scheduler
I originally thought that multiple reactors would be accessing each VdevRaid
at the same time, so the I/O scheduler was designed to enforce strict LBA
ordering of write operations.  I now intend to lock each VdevRaid to a
single reactor, so I don't need that feature.  However, it's about time that
ArkFS learned to manage its queue depth.  That's what this change does.
ArkFS now uses a C-LOOK I/O scheduler for both reads and writes, with a
single priority, and a fixed queue depth of 10.  The next step is to handle
errors, especially EAGAIN, which can happen if the process or system-wide
AIO limits are reached.