Persist the FreeSpaceMap in the Cluster label.
This is not ideal, because it requires writing the entire thing at every
sync.  However, cbor seems to do a decent job of compressing repeated byte
sequences, so perhaps it won't be too bad.  It also effectively mirrors the
FSM across all disks, which is more problematic for large arrays.  A better
approach would be to always make Zone 0 a mirrored zone (with k < n, eg
using PRIMES-7,2,2 if the default is PRIMES-7,5,2) and only write the dirty
portions of the FSM at each txg sync.