Fix a race condition in Tree::flush
Serde's serialize method can't be used in Tokio if it needs to acquire a
lock.  My version of tree_root_serializer::serialize returns EAGAIN if the
root node's lock is not immediately available.  This is inadequate for
multithreaded environments.  Fix Tree::flush by converting a locked Inner to
an InnerOnDisk, which can be directly serialized without acquiring any
locks.

Deserialize presents no race condition, because at that point the Tree is
uniquely owned.