Database::write_label must serialize _after_ syncing
Database::write_label was flushing the forest, serializing, and writing the
label in that order.  This caused future crashes, apparently due to bugs in
the allocator.  But the real reason is because when the Database's label was
serialized without flushing, the result was a label that couldn't be
correctly opened.  Opening it would result in a misaligned raid
StripeBuffer.

Fix the bug by flushing the Forest _before_ syncing.  Also, add an assertion
in StripeBuffer::new that would've caught this bug much earlier.