Tree::range: fix another panic relating to a backwards range.
This is related to change 6299ccd91470, but a little different.  Due to the
minimum key rule, it's legal for a BTree node's first key to be higher than
the key in its parent TreePtr.  Change 6299ccd91470 didn't account for that.
This situation can arise if the range's end bound lies above the parent
TreePtr's key but below the node's first key.