Tree::range_delete: fix a resource leak
range_delete_pass1 was simply dropping nodes that should be deleted
wholesale.  That works for in-memory nodes, but for on-disk nodes it caused
a leak of storage space.  The symptom was that cleaning a zone would leave a
small amount of storage still allocated in that zone.  Fix by ensuring that
range_delete_pass1 calls DML::pop or DML::delete on every node that it
deletes.