Tree::range_delete_pass1: delete nodes with 0 children
The old code correctly deleted nodes with 0 children on the left side or in
the middle of the cut, but didn't do so for nodes on the right.  Since
IntNodes only store their children's minimum keys, we don't know whether
we'll need to delete a node on the right until after recursing into it.