Remove an incorrect assertion in VdevBlock::open_zone
The assertion checks that the scheduler did not issue an I/O to a zone
immediately before opening, assuming that it would be empty.  However, that
may not always be the case.  For example, when cleaning a zone, the
scheduler will move records out of that zone and into memory.  But it might
not commit them to disk before reopening the same zone.