Rewrite Tree::range_delete_pass2 to use a bottom-up approach
This is my third version of range_delete_pass2.  This version is much
simpler, because it doesn't need to guess which nodes to fix up just because
they might underflow after their children are fixed.  For the same reason,
it does less I/O.

This commit also fixes a minor bug in TreeWriteGuard::xlock_nc: it was
wrongly setting the Node's end txg (though the next tree flush would fix
it).

Fixes bug #4a99d7f