Fix the "subtract with overflow during range_delete" bug.
The problem was that the node on the right side of the cut was being merged
during descent, and both nodes were being merged during ascent.  But
range_delete_pass2_r didn't correctly account for all the merging
operations, and thought that the right node now had an index of -1.  The
solution is to make a distinction between nodes merged to the right of the
right child in the cut, and whether the left and right children were merged
together.