Fix yet another range_delete_bug: bug 40b01eb
This bug caused a "nchildren < 2" panic.  The problem was that
range_delete_pass2_r was trying to fix a node during ascent that had no
siblings.  Such a thing is impossible.  The correct solution is to wait; the
parent will be merged higher up the stack.  In this case, the parent will be
the root, which will get merged down.