Rererewrite Tree::range_delete_pass2 to be strictly top-down.
I found a situation (bug #3a1c768) that could not be solved with the
bottom-up approach.  So this change reverts the big idea behind
790:a6967cad3f4c.  Once again, range_delete_pass2 is strictly top-down,
though it retains many changes that I made to it in the meantime.

common::fs::torture::random_clean_zone ran for 147 repetitions without a
failure.  However, I believe that there is still scope for failure when the
tree's fanout limits are narrower.