Fix a range_delete bug regarding stealing keys
For Trees with narrowly constrained node sizes (eg 3,7) it was possible that
a node in danger of underflowing would need to be fixed, but couldn't be
merged, and the old steal algorithm would give it 0 keys.  For example, in a
3,7 tree if all siblings had size 4.  That could result in underfull nodes
following range_delete.  Fix it be ensuring that at least 1 key is always
stolen at a time.

This slightly increases the amount of churn that happens during
range_delete.