Fix a range_delete underflow bug with narrowly constrained trees
I previously believed that, following range_delete_pass1, any node would
need to be fixed at most twice.  However, that assumption was incorrect with
narrowly constrained trees.  For example, in a tree with 3,7 fanout limits
there might be a child with 1 node, a sibling with 2, and another sibling
with 5.  Fixing the first node once will merge it, yielding a child with 3
children.  Fixing it a second time will steal one node from its sibling,
resulting in it having 4 children.  Since it's now a common ancestor of the
cut, it might lose two children later on, causing an underflow.

The solution is to use a loop, and fix as many times as necessary.