Split bfffs into subcrates
"bfffs" contains the library and the "bfffs" binary.
"bfffs-fuse" contains bfffsd.
Soon there will be a "bfffs-fio" containing a fio engine.
Later there will be a "bfffs-kmod" containing the bfffs kernel module.
And of course, isa-l is still with us.

The top level crate is a virtual package that will never be published.

This change also removes bfffsd and fuse/fs.rs from coverage checking.