fio needs two invocations to write and read without extra work
Unlike iozone, fio has no concept of write-then-read.  The old bfffs.fio
file was interpreted as "Do two processes simultaneously; one write test,
and one read of a nonexistent file test".  What we should do is invoke fio
once to do the write test, then invoke it a second time on the same file for
the read test.  That would also have a happy side effect of clearing the
cache.