Strip down InnerOnDisk
There's no need for InnerOnDisk to be able to represent an in-memory root
node.  That's only there to support unit tests that store Trees as strings.
Instead, create a separate InnerLiteral structure.  This change shrinks the
size of InnerOnDisk from 68 to 48 bytes.  And it also enables removing two of
InnerOnDisk's generic types.