Add a sequentially_optimized Tree parameter
For such Trees, splits will make the left node as big as possible and the
right as small as possible.  Also, fanout parameters are tweaked to account
for this.

*All* RIDT inserts are purely sequential (though overwrites may be random).
And most AllocT inserts are sequential too.  So those trees will get this
new optimization.

According to examples/fanout, the AllocT's padding is now < 1%.  The RIDT is
about 2% for random workloads, because its interior nodes can be larger than
its leaves.