Don't store the compression function in the DRP
BLOSC stores detailed information about the compression function used in the
compressed buffer itself.  The DRP only needs to use 1 bit to store whether
or not the record is compressed.  This change saves a few bytes of RAM per
DRP in the RIDT.  It also allows for the Compression enum to grow more
complicated without worrying about wasting more DRP space.