Drastically simplify the fs::dump test
The FS Tree's fanout has grown too large to for the dump test to practically
include an indirect node.  But it doesn't really need to anyway.  All of the
hard work of dump is done by the Tree code, and can be covered by its tests.
This commit reduces the dump test to an empty tree.