VdevBlock compiles now.
But I'm not happy about it!  Using owned buffers in tokio_file is ugly.
It would be more elegant if I could use borrowed buffers there.  It
would require VdevBlock to store its VdevLeaf in an Rc, and only call
tokio_file::{method} from within the task::spawn_local block.  Handling
EAGAIN would be the biggest difference.  It would have to happen
synchronously with tokio_file::{method}, instead of when the method is
polled.