Eliminate the VdevLeafApi trait
It never served any real purpose except for documentation.  But it's
causing problems as I try to move to std::future.  Some things in the
std::future world aren't compatible with dynamic dispatch.  So it's got
to go.

The Vdev trait will remain for now, because it's got two implementation:
VdevRaid and VdevOneDisk.  But it will go away in the future, too.