Flush the Database at 0.1s intervals.
Previously the Database would only be flushed when syncing a txg,
leading to large amounts of dirty data in the Tree.  Now, if the tree is
dirty, it will be flushed every 0.1s.  This is much simpler than trying
to track which key ranges need to be flushed.  However, once multiple
datasets are supported, it might be more efficient to at least track
which datasets are dirty.

Fixes #37