Fix a tree locking race condition.
Functions like Tree::xlock_root would exclusively lock a node or the
tree root, then COW one of its children.  If the child was not yet
in-memory, they would pop it from the dml and transform it into an
in-memory Node.  However, it was possible that the child was in the
cache and already locked, in which case the Arc::try_unwrap would panic.

The solution is to xlock the child, drop the guard, and then unwrap the
Arc, all while keeping the parent xlocked.  That's similar to what
Tree::flush already does.