Add tracing support
The tracing crate is a replacement for log, intended for asynchronous
programs.  This commit introduces it, just for a single event at first.
But its use will grow over time.

Note that the #[instrument] macro works best with async functions, of
which bfffs has few.  When used with a normal function that returns a
future, the future must have .in_current_span() appended.