Rename the bfffs crate to bfffs-core
End users will never use this stuff directly, so it should have the
less-memorable name.  More importantly, in the future it should have a
no_std feature, allowing it to be compiled for the kernel.  That means
that its dependent crates elsewhere in the same workspace will depend on
it with different features set.