Rename bfffs-fuse to bfffs
This will be the main crate that users will install.  bfffs-fio will
remain in its own crate, since it has a separate license.  It is also
possible that in the future bfffsd will support multiple simultaneous
access protocols, such as FUSE and NBD.  That requires them all to share
a crate.