Fs::{rename, unlink} fix a bug resulting in orphaned Inodes
Fs::rename was calling Fs::do_inactive to delete an existing destination
file.  But it took a shortcut and skipped creating the DyingInode that
do_inactive was expecting.  The result was an orphaned Inode.
Fs::unlink could also result in an orphaned Inode for the same reason,
if the file was already inactive at the time it was deleted.

Fixes #69