Use the Writeback module to limit the amount of dirty cached data
Unlike some other file systems, BFFFS's Cache module is purely
read-only.  Dirty cached data lives is owned by the Tree, not the Cache.
Previously there was no limit to it, so it was possible for bfffsd to
get ENOMEM and crash if the user wrote faster than the disks could
flush.

Now Tree's consumers must acquire Writeback Credit before they can write
to Tree.  The actual credit acquisition happens as early as possible,
before the Fs locks a txg or copies any user-supplied buffers.  Credit
is only repaid when the data gets flushed or dropped.

So far, there is not any way to set the limit on the quantity of cached
dirty data.  The "limit" is always practically infinity.

The RIDT and AllocT are special.  They don't require any Credit, since:
* They're tiny compared to the Fs Trees, and
* Flushing an Fs Tree involves writing to the RIDT and AllocT.  So we
  don't want them to block while waiting for Credit; that could cause a
  deadlock.