"Fix" a clippy redundant_allocation warning
The latest clippy warns about types like Box<Arc<T>>, because the Arc<T>
is already on the heap, so Clippy thinks that the Box has no purpose.
But in this case it does; it serves to turn the Arc<T> into a trait
object, for a trait that is implemented on the Arc<T> itself rather
than on T.  So the double allocation really is necessary.

Despite this, I "fixed" the warning basically by moving the Box's deref
from TreeReadGuard::deref to IntElem::rlock.  I think the total number
of Box deref operations will be unchanged, at least for common
operations like Tree::get.

See also https://github.com/rust-lang/rust-clippy/issues/7472