Add a Controller layer on top of Database
This is just a shim now, but it will grow in time.  It will be the entry
point into bfffs_core for all control-path operations.  That will
include disk replacement (so it must contain the dev_manager), and
multiple pool management.  For now its main function is to hide from
bfffs parts of the Database that shouldn't be exposed.