Better type safety for Fs::FileData
The FileData object must be destroyed when passed to Fs::inactive.  So
it should not implement Clone.  But it must implement Clone so that
Bfffsd::Fs's files cache's mutex doesn't need to be held across I/O.

The solution is to split it into two types: FileDataMut is not Clone,
but it can produce a FileData that is.  Most Fs methods take a FileData
argument, but Fs::inactive needs the FileDataMut.