Eliminate Dataset::remove_blob
The Fs needn't worry about manipulating blobs now.  Both writing them
and removing them is entirely the responsibility of the Tree and IDML.
It still needs to read from them, though.

This was accomplished by adding a Value::dpop method, which is basically
the opposite of Value::flush.  When removing a Value from a Tree, dpop
ensures that it pops any blobs from the IDML at the same time.