Fix pool.allocated_space
allocated_space is used primarily by the Pool in order to choose which
cluster to direct writes to.  It was always supposed to include all
allocated blocks as well as freed but not erased blocks, but never did.
Nor did it properly account for blocks wasted when closing a zone or
syncing a txg.  Fix it by:

* Moving the calculations from Pool down into Cluster, where the stuff
  actually happens.  This also makes the unit testing easier, and saves
  one division operation on every write.

* Propertly accounting for freeing, erasing, finishing, and syncing
  operations.

Fixes #180