Update fio bindings for fio-3.30
bindgen produces the wrong struct layout for thread_data.  It calculates
the wrong offset for client_type.  Workaround the bug by binding the
offending type as opaque.  In that case, bindgen calculates its size
correctly.

Include the layout tests after all, to help catch bugs like this.

The layout bug was previously just an annoyance, but with fio-3.30 it
causes failures at runtime.

https://github.com/rust-lang/rust-bindgen/issues/2101