Improve FS tree deserialize performance
Dtrace shows that the cpu consumption of "fs destroy" is dominated by
memory allocations.  FSValue::Inode is almost three times larger than
the next largest variant, so box it, reducing the in-memory size of
FSValue, and reducing the amount of memory allocator activity when
working with large files.

Improves the throughput of Node::deserialize for an FS leaf node by
about 50%.