VdevFile: ensure all writes are block-sized and block-aligned
The OS will reject with EINVAL any write that isn't block-sized and
block-aligned.  That includes aio_writev operations where the operation
is block-aligned and block-sized overall, but one of the individual
iovecs is not.

Previously tokio-file took care of this, but the next version will not.
It's ugly, and it actually involves more data copies than doing it in
BFFFS.