move vdev_file's copying and zero-padding into vdev_block
The OS requires aio_writev and aio_readv sglists to be multiples of a
block size, not just for the operation overall but for each iovec.
BFFFS's DDML takes care of that for reads.  For writes, previously
vdev_file did it.  But now vdev_block must, so that it can correctly
accumulate adjacent operations.