Eliminate duplicate disk reads coming from the DMLs
*DML::get would first check the cache.  If the requested block weren't
there, then it would read from disk and insert the result into the
cache.  However, it is possible that many tasks simultaneously attempt
to read the same block from cache.  When that happened the result would
be a flood of disks reads reading the same exact LBA over and over.

Fix this performance bug by adding a queue of tasks that are waiting on
the same cache insertion.  Speeds up fs_destroy by about 4.5x .

Fixes #198