Use EV_ONESHOT with posix AIO
When Tokio gets event notification from kevent, the pending future may
belong to a different thread.  If so, it signals the other thread and
pends on kevent again.  But the other thread may not have had time to
call aio_return yet.  In that case, the first thread's kevent will
immediately return again.

The solution is to set EV_ONESHOT on the aiocb.  Nix 0.27.0 includes
that change, along with mio-aio 0.8.0 and tokio-file 0.9.0.

Using EV_ONESHOT allows us to use a multi-threaded Tokio runtime, which
improves performance by 2.8x.