Make Value::flush operate on a BTreeMap at a time
Primary by passing over Values that don't need to be flushed, it greatly
reduces malloc activity.  It improves the performance of a 64-file,
1GiB/file write benchmark from 396 to 875 MB/s.

Operating on a whole BTreMap at a time is less obvious than operating on
a Value at a time.  But it allows eliminating more allocations while
requiring less code than any alternatives.