Implement mirroring
Mirrors like above VdevBlock but below Raid.  This makes it possible to
create something like a RAID510 pool, though I don't know why you'd want
to.  The reason to do it this way is because in the future the Mirror
layer will be exploited for disk replacements.

Currently only healthy configurations are supported.

Example RAID510 pool:

testpool
    raid 3, 1
        mirror
	    da0
	    da1
	mirror
	    da2
	    da3
	mirror
	    da4
	    da5
    raid 3, 1
        mirror
	    da6
	    da7
	mirror
	    da8
	    da9
	mirror
	    da10
	    da11

Also, reverse the API for creating pools.

Previously it was a top-down API: Pool::create calls Cluster::create and
so on down the chain.  But with the introduction of mirroring, that's
too complicated.  Instead use a bottom-up API: Pool::create takes an
already created Cluster, Cluster::create takes an already created Raid,
etc etc.

Also, factor out common code in the functional tests.

Add a PoolBuilder for use in setup functions