Use I/O safety in vdev_file
Update to a new version of nix and tokio-file's that enforce I/O safety.
But because vdev_block includes a scheduler which stores incomplete
futures in a collection, the futures must be 'static.  Rather than shove
an Arc<File> into each one, we rely on vdev_block's drop order to know
that all futures will be complete before dropping the File.  So we don't
rely on tokio-file's's I/O Safety guarantees.  Instead, we unsafely
bypass them.

This change also moves vdev_file's Manager struct up into vdev_block,
because the Manager doesn't need to know the nitty-gritty details of how
vdevs are implemented.