Account for allocated space in credit_requirements()
In addition to the memory consumption of a tree node's keys and values
themselves, some Value types allocate additional space, like
InlineExtent.  We must account for that in Tree::credit_requirements.
This fixes a credit underflow bug when working with trees containing
many small files.

However, the fix is incomplete.  It isn't possible to guarantee any
particular upper bound on allocated space, because nothing within BFFFS
limits the size of file names or extended attribute names.

This commit also relies on nightly-only features, so can't be merged
as-is.