Delay the erase and reuse of freed zones
After freeing the last lba from a zone, it must not be erased or reused
until after the start of the next transaction, so the file system may be
rolled back if it crashes before the next transaction sync.

Instead, mark such zones as "dead", and erase them at the end of the
current transaction.

Fixes #261