Fix Database::check with dirty trees
It always worked for a pool where everything was clean.  But if any Tree
was dirty, the operation would fail with ENOENT.

The problem was that the Forest stores the serialized head of each tree,
_even_ if that tree is dirty.  But if dirty, then its root node won't
actually be on disk.  That is, when a Tree is dirty, the Forest stores
an out-of-date head.  That's by-design; the database isn't supposed to
access the Forest for dirty trees.  But the Database's check method did
anyway.  Fix the bug in Database::check.