Eliminate the DevManager
It as an enormous layering violation, for one thing.  For another, it
isn't a useful abstraction for the purpose of replacing failed disks,
because it doesn't retain any kind of references to the VdevRaid and
Mirror objects.

This is a partial reversion of 99022df1e4743b9abffa6a8804989944b03c15dc.
They key difference between this code and the code from before that
revision is that now there is a Manager struct at each layer.  That
allows for tasting and sparing, but doesn't violate layering like
DevManager did.

Also, eliminate the ugly DeviceManager::import_clusters method by adding
a Database::dump_fsm.  It's much nicer.