Make use of Manager in functional tests
This is a much more more straightforward way to import vdevs than the
old way, which basically reproduced the pool import logic in each test
file.  It's a nice spinoff from eliminating the DevManager.

Also, fix some inconsistencies in the Cluster golden label that
previously weren't apparent, because we didn't verify it.  We only
really verified the golden spacemap.