Switch from serde_yaml to serde_yaml_ng
Because serde_yaml is no longer maintained, and the version we were
using also depended on an abandoned crate (yaml-rust). serde_yaml 0.9
(and by extention serde_yaml_ng) changed the serialization format for
enums, which makes bfffs tree dumps more compact.  This changes the
output format of the "bfffs debug dump" command.  But it does not change
the on-disk format.