Update lalrpop and regex
To eliminate duplicate dependency on regex-syntax, and fix a lint in
lalrpop's generated code.

Also update prost-derive and predicates, to mostly remove a duplicate
dependency on itertools created by the lalrpop upgrade.  Can't
completely remove it until the next release of criterion.  cargo-bloat
shows that regex-syntax is much larger than itertools, so this is still
a net win.