Replace bincode with speedy
Because bincode is no longer maintained.  This commit derives speedy for
all serializable types, and switches the labels and tree to use it.

This also removes Serialize/Deserialize in many places.  But not all,
because we still need it to dump a Tree as YAML.

In terms of performance, the new code has about 10% higher interior Tree
fanout, because of more compact encoding.  But Tree Leaf nodes are
unchanged.

It's also a little bit faster, probably because Speedy bypasses the
Serde intermediate layer.  Compared to bincode, Speedy is:

alloct-leaf/serialize:		101.31% as fast
alloct-leaf/deserialize:	293.18% as fast
fs-leaf/serialize:		96.24% as fast
fs-leaf/deserialize:		150.09% as fast
ridt-leaf/serialize:		167.77% as fast
ridt-leaf/deserialize:		125.56% as fast